pyproject.toml
[project]
name = "multi-agent-reviewer"
version = "0.1.0"
description = "Multi-agent peer review system for biomedical manuscripts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.40.0",
    "pypaperretriever @ git+https://github.com/josephisaacturner/pypaperretriever.git",
    "pymupdf>=1.24.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "anthropic>=0.75.0",
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
reviewer = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

